requirements.txt
absl-py==1.4.0
astor==0.8.1
audioread==3.0.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
Cython==0.29.33
dataclasses==0.6
decorator==5.1.1
Distance==0.1.3
future==0.18.3
g2p-en==2.1.0
gast==0.2.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.2
h5py==3.1.0
idna==3.4
importlib-metadata==4.8.3
inflect==4.1.0
joblib==1.1.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
librosa==0.7.2
llvmlite==0.31.0
Markdown==3.3.7
matplotlib==3.2.2
nltk==3.6.7
numba==0.48.0
numpy==1.19.0
oauthlib==3.2.2
opt-einsum==3.3.0
Pillow==8.4.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.7
pypinyin==0.39.0
python-dateutil==2.8.2
pyworld==0.2.10
PyYAML==5.4.1
regex==2022.10.31
requests==2.27.1
requests-oauthlib==1.3.1
resampy==0.3.1
rsa==4.9
scikit-learn==0.23.2
scipy==1.5.0
six==1.16.0
SoundFile==0.10.3.post1
tensorboard==1.15.0
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
tgt==1.4.4
threadpoolctl==3.1.0
tqdm==4.46.1
typing_extensions==4.1.1
Unidecode==1.1.1
urllib3==1.26.15
Werkzeug==2.0.3
wrapt==1.15.0
zipp==3.6.0
